Done all for logs. Time to write about my work

+++ app/src/main/java/com/faa/knowyourgame_new/entity/Logs.java
@@ -23,7 +23,7 @@ public class Logs {
     int answerStatus;
 
     @ColumnInfo(name = "points")
-    double points;
+    int points;
 
     @ColumnInfo(name = "date_time")
     String dateTime;
